feat: add support for building server and client targets in xmake and update README for client library usage

+++ xmake.lua
@@ -8,6 +8,18 @@ option("with-autoupdate")
     set_description("Enable automatic game download/update feature")
 option_end()
 
+option("pack-server")
+    set_default(true)
+    set_showmenu(true)
+    set_description("Include server in package")
+option_end()
+
+option("pack-client")
+    set_default(true)
+    set_showmenu(true)
+    set_description("Include client library in package")
+option_end()
+
 add_requires("nlohmann_json", "singleton")
 
 if has_config("with-autoupdate") then
@@ -20,7 +32,7 @@ includes("@builtin/xpack")
 set_project("Flakkari")
 set_license("MIT")
 
-target("flakkari")
+target("flakkari-server")
     set_kind("binary")
     set_default(true)
     set_languages("cxx20")
@@ -46,12 +58,17 @@ target("flakkari")
 
     add_files("Flakkari/**.cpp")
 
+    remove_files("Flakkari/Client/**.cpp")
+
     if not has_config("with-autoupdate") then
         remove_files("Flakkari/Server/Internals/GameDownloader.cpp")
     end
 
     add_headerfiles("Flakkari/**.h", { public = true })
     add_headerfiles("Flakkari/**.hpp", { public = true })
+
+    remove_headerfiles("Flakkari/Client/**.hpp")
+
     add_includedirs("Flakkari/", { public = true })
     add_includedirs("Flakkari/Engine", { public = true })
     add_includedirs("Flakkari/Engine/EntityComponentSystem", { public = true })
@@ -69,19 +86,64 @@ target("flakkari")
     if is_plat("windows") then
         add_syslinks("Iphlpapi")
     end
+target_end()
+
+target("flakkari-client")
+    set_kind("static")
+    set_default(false)
+    set_languages("cxx20")
+    set_policy("build.warning", true)
+    set_version("0.8.0")
+
+    if is_mode("debug") then
+        add_defines("_DEBUG")
+        set_symbols("debug")
+        set_optimize("none")
+    elseif is_mode("release") then
+        add_defines("NDEBUG")
+        set_optimize("fastest")
+    end
+
+    add_files("Flakkari/**.cpp")
+
+    remove_files("Flakkari/*.cpp")
+    remove_files("Flakkari/Server/**.cpp")
+
+    add_headerfiles("Flakkari/**.h", { public = true })
+    add_headerfiles("Flakkari/**.hpp", { public = true })
+
+    remove_headerfiles("Flakkari/Server/**.hpp")
+
+    add_includedirs("Flakkari/", { public = true })
+    add_includedirs("Flakkari/Engine", { public = true })
+    add_includedirs("Flakkari/Engine/EntityComponentSystem", { public = true })
+    add_includedirs("Flakkari/Engine/EntityComponentSystem/Components", { public = true })
+    add_includedirs("Flakkari/Engine/EntityComponentSystem/Systems", { public = true })
+    add_includedirs("Flakkari/Engine/Math", { public = true })
+    add_includedirs("Flakkari/Logger", { public = true })
+    add_includedirs("Flakkari/Network", { public = true })
+    add_includedirs("Flakkari/Protocol", { public = true })
+target_end()
 
 -- Doc: https://xmake.io/#/manual/xpack
 xpack("flakkari")
     set_formats("nsis", "zip", "srczip", "targz", "srctargz", "runself")
-    set_title("Flakkari Server v$(version) ($(arch))")
+    set_title("Flakkari Full v$(version) ($(arch))")
     set_author("MasterLaplace")
     set_company("ME.inc")
-    set_description("Flakkari is a UDP/TCP game server initially developed for the R-Type Epitech project and updated for the Video Games course at University Laval.")
+    set_description("Flakkari is a UDP/TCP game server and client library initially developed for the R-Type Epitech project and updated for the Video Games course at University Laval.")
     set_homepage("https://github.com/MasterLaplace/Flakkari")
     set_licensefile("LICENSE")
     set_copyright("Copyright (C) 2023-present, Me.inc & MasterLaplace")
     set_version("0.8.0")
-    add_targets("flakkari")
+
+    if has_config("pack-server") then
+        add_targets("flakkari-server")
+    end
+    if has_config("pack-client") then
+        add_targets("flakkari-client")
+    end
+
     set_iconfile("docs/Images/icon.ico")
 
     add_sourcefiles("Flakkari/**.cpp")
@@ -93,10 +155,22 @@ xpack("flakkari")
     add_installfiles("*.md", {prefixdir = "share/docs"})
 
     on_load(function (package)
+        local pack_server = has_config("pack-server")
+        local pack_client = has_config("pack-client")
+        local pack_type = ""
+
+        if pack_server and pack_client then
+            pack_type = "full"
+        elseif pack_server then
+            pack_type = "server"
+        elseif pack_client then
+            pack_type = "client"
+        end
+
         if package:from_source() then
-            package:set("basename", "flakkari-$(plat)-src-v$(version)")
+            package:set("basename", "flakkari-" .. pack_type .. "-$(plat)-src-v$(version)")
         else
-            package:set("basename", "flakkari-$(plat)-$(arch)-v$(version)")
+            package:set("basename", "flakkari-" .. pack_type .. "-$(plat)-$(arch)-v$(version)")
         end
 
         local format = package:format()
@@ -108,6 +182,10 @@ xpack("flakkari")
     end)
 
     after_installcmd(function (package, batchcmds)
+        if not has_config("pack-server") then
+            return
+        end
+
         batchcmds:mkdir(package:installdir("Games"))
         batchcmds:cp("Games/*", package:installdir("Games"), {rootdir = "."})
 
@@ -127,6 +205,9 @@ xpack("flakkari")
     end)
 
     after_uninstallcmd(function (package, batchcmds)
+        if not has_config("pack-server") then
+            return
+        end
         batchcmds:rmdir(package:installdir("Games"))
 
         if is_plat("windows") then
@@ -141,3 +222,4 @@ xpack("flakkari")
             batchcmds:showmsg("Please remove the environment variable FLAKKARI_INSTALL_GAME_DIR manually if it was set.")
         end
     end)
+xpack_end()
